Fixing some sound interpolation issue in nec cpu

+++ src/cpu/nec_intf.cpp
@@ -211,8 +211,8 @@ UINT32 VezReadLong(UINT32 a)
 	if ( p )
 		return *(p + (a / 4));
 	else
-		return VezCurrentCPU->ReadHandler(a) + (VezCurrentCPU->ReadHandler(a+1) * 0x100) + 
-			(VezCurrentCPU->ReadHandler(a+2) * 0x10000) + (VezCurrentCPU->ReadHandler(a+3) * 0x1000000);
+		return ((UINT32)VezCurrentCPU->ReadHandler(a) + (VezCurrentCPU->ReadHandler(a+1) * 0x100) + 
+			(VezCurrentCPU->ReadHandler(a+2) * 0x10000) + (UINT32)(VezCurrentCPU->ReadHandler(a+3) * 0x1000000));
 }
 
 static void VezCheatWrite(UINT32 a, UINT8 d)
